Cast return type of Class.new using RBI sigs
TODO: building in release mode and running on Shopify core segfaults

The goal of this commit is to cast the return of Class.new so that it
takes into account the type of the argument. This is our third attempt,
this time using RBI sigs and generics to specify that the return type
should be the same as the argument.